renaming all "class" files to UpperCamelCase.
