+++ index.js
@@ -1,21 +1,21 @@
 if (module && require && require.main == module) {
   var json = require('./diagram.json');
-  var JSONFactory = require('./lib/factories/json');
-  var GraphvizDot = require('./lib/exporters/graphvizdot');
+  var JSONFactory = require('./lib/factories/JSONFactory');
+  var GraphvizDot = require('./lib/exporters/GraphvizDot');
 
   var doc = (new JSONFactory()).build(json);
   var dot = new GraphvizDot(doc);
 
   console.log(dot.export());
 } else {
-  exports.Document = require('./lib/document');
-  exports.ClassDiagram = require('./lib/diagrams/classdiagram');
-  exports.Class = require('./lib/elements/class');
-  exports.Property = require('./lib/elements/property');
-  exports.Method = require('./lib/elements/method');
-  exports.Link = require('./lib/elements/link');
-  exports.GraphvizDot = require('./lib/exporters/graphvizdot');
-  exports.JSONFactory = require('./lib/factories/json');
-  exports.Variable = require('./lib/elements/variable');
+  exports.Document = require('./lib/Document');
+  exports.ClassDiagram = require('./lib/diagrams/ClassDiagram');
+  exports.Class = require('./lib/elements/Class');
+  exports.Property = require('./lib/elements/Property');
+  exports.Method = require('./lib/elements/Method');
+  exports.Link = require('./lib/elements/Link');
+  exports.GraphvizDot = require('./lib/exporters/GraphvizDot');
+  exports.JSONFactory = require('./lib/factories/JSONFactory');
+  exports.Variable = require('./lib/elements/Variable');
   exports.Visibility = require('./lib/shared/visibility');
 }
