+++ lib/Document.js
