+++ lib/diagrams/ClassDiagram.js
