+++ lib/elements/Class.js
