+++ lib/elements/Method.js
@@ -1,6 +1,6 @@
 var util = require('util');
 var _ = require('lodash');
-var Property = require('./property');
+var Property = require('./Property');
 
 function Method(visibility, variable, args) {
   Method.super_.call(this, visibility, variable);
