+++ lib/elements/Property.js
