+++ lib/exporters/GraphvizDot.js
@@ -1,7 +1,7 @@
 var util = require('util');
 var _ = require('lodash');
 
-var Link = require('../elements/link');
+var Link = require('../elements/Link');
 
 function GraphvizDot(document) {
   this.document = document;
