+++ lib/factories/JSONFactory.js
@@ -1,15 +1,15 @@
 var util = require('util');
 var _ = require('lodash');
 
-var Document = require('../document');
+var Document = require('../Document');
 
-var ClassDiagram = require('../diagrams/classdiagram');
+var ClassDiagram = require('../diagrams/ClassDiagram');
 
-var Class = require('../elements/class');
-var Property = require('../elements/property');
+var Class = require('../elements/Class');
+var Property = require('../elements/Property');
 var Method = require('../elements/Method');
-var Link = require('../elements/link');
-var Variable = require('../elements/variable');
+var Link = require('../elements/Link');
+var Variable = require('../elements/Variable');
 
 function JSONFactory() {
   this.refs = {};
